typify the ProfilePage
