+++ src/types/Geojson.ts
@@ -1,0 +1,8 @@
+import Feature from "types/Feature";
+
+type Geojson = {
+  type: "FeatureCollection";
+  features: Feature[];
+};
+
+export default Geojson;
