+++ src/types/Photo.ts
@@ -3,6 +3,9 @@ type Photo = {
   main: any;
   thumbnail: any;
   updated: any;
+  moderated: any;
+  owner_id: string;
+  pieces: number;
 };
 
 export default Photo;
